Split import_xml_doc into several modules.
